// src/app/_models/dto/location-result-dto.ts
export class LocationResultDTO {
    id: number;
    storefrontId: number;
    address1: string;
    address2: string;
    city: string;
    displayName: string;
    distance: number;
    imgUrl: string;
    latitude: number;
    longitude: number;
    rating: number;
    reviewCount: number;
    state: string;
    zip: string;
}